:truck: Moved `Primitive` to `PrimitiveGen`
